service/requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.*
google-cloud-logging==3.*
google-cloud-firestore==2.*
firebase-admin==6.*
google-auth==2.*
google-auth-oauthlib==1.*
google-auth-httplib2==0.*
google-api-python-client==2.*
langchain-core
langchain-openai
openai
PyYAML
